Test: Improve description and redact name test.
The testing for name change in Question is not meaningful, unlike how it
is meaningful in User, which is actually testing a method named name_was,
and not a mere activerecord name attribute. Remove the useless test from
Question, and improve the description of the name_was test in User.