Question: Test that we cannot delete if it has answers.
An Answer belongs_to a Question, and has a foreign key constraint in the schema.
Deleting a Question that has answers would result in orphaned answers. The FK constraint
is enforcing this at the db schema level.

Test to prove that we cannot delete a Question that has associated Answers.

References #29.