Scholarship: Cascade deletion of questions.
This adds a cascade deletion in the schema, such that deleting a Scholarship
deletes any associated Questions. It also declares a cascading delete in the
model association.

Note that, because a Question cannot be destroyed if it has Answers (see #29),
we cannot destroy a Scholarship that has a Question that has any Answers. This
is the intended behavior.

References #34.