Scholarship: Refine deadline validations.
The existing validator ensures a scholarship's application_deadline cannot be
in the past. But, there could be cases where a change is made to a scholarship
after the deadline.

Refine the existing application_deadline validation to ensure that the
date cannot be in the past when created only.
Add a new validator that ensures an existing scholarship with a deadline
in the past can be updated, but only if the application_deadline isn't
changed.