User: Rename applied_for to started_application_for.
We will need a simple way to discern between users who have begun an application
and those who have submitted an application for a scholarship.

Rename the method applied_for to just started_application_for, since it only
checks for the presence of an application associated with the user. The phrase
'applied for' carries too much impllying that the application has been submitted.