Admin Applications: Add preview feature.
Let's leverage new for the request semantics, since it's job is to display
the application form. Duplicate the applicant's 'new' view and enable the
route and controller action. Disable the view's submit button, and add
a flash message that states the admin is viewing a preview, with a link
to go back to the scholarship show view.

References #65.