Application: Enable recommender, as optional.
The class has its belongs_to :recommender declaration commented out,
along with some recommender-related code in to_s commented out.

Uncomment this recommender code, clean up, and set to optional. The
recommender is optional because we need to be able to save the
scholarship without a recommender in the workflow.

References #43.