Rails 7 Upgrade: Update delete links and controller actions.
Rails 7 does not pay attention to the data-method attribute of links any
more, because that is a UJS thing. Turbo needs to see the data-turbo-method
attribute in order to hijack the link and send a DELETE request.

Add the data-turb-method and data-turbo-confirm attributes to restore the
prior behavior. Keep the method: :delete attribute there, because Capybara
is not yet updated to be aware of the new turbo attributes. Without it,
Capybara's Rack Driver won't send the right request type when clicking on
delete links.

In addition, Turbo-generated DELETE requests are more finicky. The redirect
instruction they receive will use the SAME request type, and therefore
try to send, say, a DELETE request to an index view after a delete action
completes. This isn't valid.

So, ensure that controller delete actions send a 303 redirect, informing
the client that the redirect request should NOT use the request type of
the original request, and instead should just issue a GET.

References #141.